+++ src/readline/terminal.h
@@ -28,7 +28,7 @@ struct terminal
     s_string *input;
     size_t input_index;
     s_hist_entry *hist_current;
-    size_t hist_pos;
+    int hist_pos;
     char *prompt;
 };
 typedef struct terminal s_term;
